Move Modal and ModalImpl into their own packages
